requirements.txt
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
configparser==6.0.1
DateTime==5.5
dnspython==2.6.1
filelock==3.13.3
future==1.0.0
humanfriendly==10.0
idna==3.6
nslookup==1.7.0
ping3==4.0.8
PyJWT==2.8.0
pyping==0.0.6
pyserial==3.5
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
setuptools==69.2.0
urllib3==2.2.1
uuid==1.30
webex-bot==0.4.1
webexteamssdk==1.6.1
websockets==10.2
zope.interface==6.2
